Update bikeshare.py
Missing functionality:

There is a missing functionality! The user must be able to ask you to show some rows of data upon request.

Please create a new function for this, for example called display_data. You need to ask the user whether he wants to see 5 rows of data. We will show all parts of analysis then we will ask this question: Do you want to see 5 rows of data. But it is not only the first 5 data but as much as the user wants. For example, you asked the user like this in the first step: "Do you want to see the first 5 rows of data?". He typed yes and you will show the first 5 rows ,then you ask again " Do you want to see the next 5 rows of data?" then he says yes and you show the next 5 data. You need to keep asking until he says no.

Here are some hints and template code:

we will show the data based on the location. namely, we will show the first 5 data in the first attempt, then second 5 data for the second "yes" So we need to keep track of this. How? (I used start_loc variable)
please check the iloc function. It returns dataframe based on the position. For example df.iloc[0:5] will return first 5 rows of data
what should be our condition for while condition? Remember, as long as the user does not enter no, we are inside in the loop.
Dont forget to ask the question inside the loop, otherwise it will be infinite loop.